refactor to share same account param editor for both mongo and cassandra API.
