+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-cosmos/src/main/java/com/microsoft/azure/toolkit/intellij/cosmos/dbtools/AzureCosmosDbAccountParamEditor.java
@@ -6,13 +6,13 @@
 package com.microsoft.azure.toolkit.intellij.cosmos.dbtools;
 
 import com.intellij.credentialStore.OneTimeString;
+import com.intellij.database.dataSource.DataSourceConfigurable;
 import com.intellij.database.dataSource.LocalDataSource;
 import com.intellij.database.dataSource.url.DataInterchange;
 import com.intellij.database.dataSource.url.FieldSize;
-import com.intellij.database.dataSource.url.TypesRegistry;
+import com.intellij.database.dataSource.url.template.UrlEditorModel;
 import com.intellij.database.dataSource.url.ui.ParamEditorBase;
-import com.intellij.openapi.ui.ComboBox;
-import com.intellij.util.Consumer;
+import com.intellij.ui.components.JBCheckBox;
 import com.microsoft.azure.toolkit.intellij.common.AzureComboBox;
 import com.microsoft.azure.toolkit.lib.Azure;
 import com.microsoft.azure.toolkit.lib.auth.AzureAccount;
@@ -21,38 +21,38 @@ import com.microsoft.azure.toolkit.lib.common.task.AzureTask;
 import com.microsoft.azure.toolkit.lib.common.task.AzureTaskManager;
 import com.microsoft.azure.toolkit.lib.cosmos.AzureCosmosService;
 import com.microsoft.azure.toolkit.lib.cosmos.CosmosDBAccount;
-import com.microsoft.azure.toolkit.lib.cosmos.cassandra.CassandraCosmosDBAccount;
-import com.microsoft.azure.toolkit.lib.cosmos.model.CassandraDatabaseAccountConnectionString;
+import com.microsoft.azure.toolkit.lib.cosmos.model.CosmosDBAccountConnectionString;
 import com.microsoft.azure.toolkit.lib.cosmos.model.DatabaseAccountKind;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.Setter;
+import lombok.SneakyThrows;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.commons.lang3.reflect.FieldUtils;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import javax.annotation.Nonnull;
 import java.util.Collections;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
-public class CassandraCosmosDbAccountParamEditor extends ParamEditorBase<ComboBox<CassandraCosmosDBAccount>> {
+public class AzureCosmosDbAccountParamEditor extends ParamEditorBase<AzureCosmosDbAccountParamEditor.CosmosDbAccountComboBox> {
     public static final String KEY_COSMOS_ACCOUNT_ID = "AZURE_COSMOS_ACCOUNT";
     public static final String KEY_USERNAME = "username";
     @Getter
     @Setter
     private String text = "";
-    private CassandraCosmosDBAccount account;
-    private CassandraDatabaseAccountConnectionString connectionString;
+    private CosmosDBAccount account;
+    private CosmosDBAccountConnectionString connectionString;
     private boolean updating;
 
-    public CassandraCosmosDbAccountParamEditor(@NotNull String label, @NotNull DataInterchange interchange) {
-        super(new CassandraCosmosDbAccountComboBox(), interchange, FieldSize.LARGE, label);
+    public AzureCosmosDbAccountParamEditor(@Nonnull DatabaseAccountKind kind, @NotNull String label, @NotNull DataInterchange interchange) {
+        super(new CosmosDbAccountComboBox(kind), interchange, FieldSize.LARGE, label);
 
-        final CassandraCosmosDbAccountComboBox combox = (CassandraCosmosDbAccountComboBox) this.getEditorComponent();
+        final CosmosDbAccountComboBox combox = this.getEditorComponent();
         interchange.addPersistentProperty(KEY_COSMOS_ACCOUNT_ID);
         interchange.addPersistentProperty(KEY_USERNAME);
         interchange.addPersistentProperty("user");
@@ -65,15 +65,15 @@ public class CassandraCosmosDbAccountParamEditor extends ParamEditorBase<ComboBo
 
     private void onPropertiesChanged(String propertyName, Object newValue) {
         if (!this.updating && Objects.nonNull(this.connectionString) && StringUtils.isNotEmpty((String) newValue)) {
-            if (StringUtils.equals(propertyName, "host") && !Objects.equals(this.connectionString.getContactPoint(), newValue) ||
+            if (StringUtils.equals(propertyName, "host") && !Objects.equals(this.connectionString.getHost(), newValue) ||
                 StringUtils.equals(propertyName, "port") && !Objects.equals(this.connectionString.getPort() + "", newValue)) {
-                ((CassandraCosmosDbAccountComboBox) this.getEditorComponent()).setValue((CassandraCosmosDBAccount) null);
+                this.getEditorComponent().setValue((CosmosDBAccount) null);
                 this.setAccount(null);
             }
         }
     }
 
-    private void setAccount(@Nullable CassandraCosmosDBAccount account) {
+    private void setAccount(@Nullable CosmosDBAccount account) {
         if (this.updating || Objects.equals(this.account, account)) {
             return;
         }
@@ -87,18 +87,17 @@ public class CassandraCosmosDbAccountParamEditor extends ParamEditorBase<ComboBo
         this.updating = true;
         // AzureActionManager.getInstance().getAction(Action.REQUIRE_AUTH).handle(combox::reloadItems);
         AzureTaskManager.getInstance().runOnPooledThread(() -> {
-            this.connectionString = account.getCassandraConnectionString();
+            this.connectionString = account.getCosmosDBAccountPrimaryConnectionString();
             final LocalDataSource dataSource = interchange.getDataSource();
-            final String host = connectionString.getContactPoint();
+            final String host = connectionString.getHost();
             final String port = String.valueOf(connectionString.getPort());
             final String user = connectionString.getUsername();
             final String password = String.valueOf(connectionString.getPassword());
-            CassandraCosmosDbAccountParamEditor.this.text = connectionString.getConnectionString();
+            this.text = connectionString.getConnectionString();
             AzureTaskManager.getInstance().runLater(() -> {
-                interchange.getDataSource().setAuthProviderId(AzurePluginAuthProvider.ID);
                 LocalDataSource.setUsername(dataSource, user);
                 interchange.getCredentials().storePassword(dataSource, new OneTimeString(password));
-                final Map<String, String> map = new HashMap<>();
+                this.setUseSsl(true);
                 interchange.putProperties(consumer -> {
                     consumer.consume(KEY_COSMOS_ACCOUNT_ID, account.getId());
                     consumer.consume(KEY_USERNAME, user);
@@ -107,35 +106,42 @@ public class CassandraCosmosDbAccountParamEditor extends ParamEditorBase<ComboBo
                     consumer.consume("port", port);
                     this.updating = false;
                 });
+                this.setUsernameAndPassword(user, password);
             }, AzureTask.Modality.ANY);
         });
     }
 
-    static class CassandraCosmosDbAccountFactory implements TypesRegistry.TypeDescriptorFactory {
-        private static final String TYPE_NAME = "cosmos_account_cassandra";
-        private static final String CAPTION = "Account";
-        private static final String PARAM_NAME = "account";
-
-        public void createTypeDescriptor(@NotNull Consumer<? super TypesRegistry.TypeDescriptor> consumer) {
-            consumer.consume(new TypesRegistry.BaseTypeDescriptor(TYPE_NAME, ".", CAPTION) {
-                @NotNull
-                protected TypesRegistry.@NotNull ParamEditor createFieldImpl(@NotNull String caption, @Nullable String configuration, @NotNull DataInterchange interchange) {
-                    return new CassandraCosmosDbAccountParamEditor(formatFieldCaption(CAPTION), interchange);
-                }
-            });
-        }
+    private void setUsernameAndPassword(String user, String password) {
+        final UrlEditorModel model = this.getDataSourceConfigurable().getUrlEditor().getEditorModel();
+        model.setParameter("user", user);
+        model.setParameter("password", password);
+        model.commit(true);
     }
 
-    private static class CassandraCosmosDbAccountComboBox extends AzureComboBox<CassandraCosmosDBAccount> {
+    @SneakyThrows
+    private void setUseSsl(boolean useSsl) {
+        final DataSourceConfigurable configurable = this.getDataSourceConfigurable();
+        final JBCheckBox useSSLCheckBox = (JBCheckBox) FieldUtils.readField(configurable.getSshSslPanel(), "myUseSSLJBCheckBox", true);
+        useSSLCheckBox.setSelected(useSsl);
+    }
+
+    @SneakyThrows
+    private DataSourceConfigurable getDataSourceConfigurable() {
+        return (DataSourceConfigurable) FieldUtils.readField(this.getInterchange(), "myConfigurable", true);
+    }
+
+    @RequiredArgsConstructor
+    static class CosmosDbAccountComboBox extends AzureComboBox<CosmosDBAccount> {
+        private final DatabaseAccountKind kind;
 
         @Nonnull
         @Override
-        protected List<CassandraCosmosDBAccount> loadItems() {
+        protected List<CosmosDBAccount> loadItems() {
             if (!Azure.az(AzureAccount.class).isLoggedIn()) {
                 return Collections.emptyList();
             }
-            return Azure.az(AzureCosmosService.class).accounts(DatabaseAccountKind.CASSANDRA).stream()
-                .filter(m -> !m.isDraftForCreating()).map(a -> ((CassandraCosmosDBAccount) a)).collect(Collectors.toList());
+            return Azure.az(AzureCosmosService.class).accounts(kind).stream()
+                .filter(m -> !m.isDraftForCreating()).map(a -> a).collect(Collectors.toList());
         }
 
         @Override
