+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-cosmos/src/main/java/com/microsoft/azure/toolkit/intellij/cosmos/dbtools/AzurePluginAuthInterceptor.java
@@ -22,7 +22,7 @@ public class AzurePluginAuthInterceptor implements DatabaseConnectionInterceptor
     public CompletionStage<ProtoConnection> intercept(@NotNull DatabaseConnectionInterceptor.ProtoConnection proto, boolean silent) {
         final DatabaseConnectionPoint connectionPoint = proto.getConnectionPoint();
         final Credentials credentials = proto.getCredentials().getCredentials(connectionPoint);
-        final String username = connectionPoint.getAdditionalProperty(MongoCosmosDbAccountParamEditor.KEY_USERNAME);
+        final String username = connectionPoint.getAdditionalProperty(AzureCosmosDbAccountParamEditor.KEY_USERNAME);
         if (StringUtils.isBlank(credentials.getUserName()) && StringUtils.isNotBlank(username)) {
             DatabaseCredentialsAuthProvider.applyCredentials(proto, new Credentials(username, credentials.getPasswordAsString()), false);
             return CompletableFuture.completedFuture(proto);
