+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-cosmos/src/main/resources/META-INF/azure-intellij-plugin-cosmos-database.xml
@@ -10,7 +10,7 @@
         <extensionFallback dbms="AZ_COSMOS_CASSANDRA" fallbackDbms="CASSANDRA"/>
         <driversConfig config="databaseDrivers/azure-cosmos-mongo-drivers.xml"/>
         <driversConfig config="databaseDrivers/azure-cosmos-cassandra-drivers.xml"/>
-        <urlParamEditorProvider implementation="com.microsoft.azure.toolkit.intellij.cosmos.dbtools.MongoCosmosDbAccountParamEditor$MongoCosmosDbAccountFactory" order="first"/>
-        <urlParamEditorProvider implementation="com.microsoft.azure.toolkit.intellij.cosmos.dbtools.CassandraCosmosDbAccountParamEditor$CassandraCosmosDbAccountFactory" order="first"/>
+        <urlParamEditorProvider implementation="com.microsoft.azure.toolkit.intellij.cosmos.dbtools.MongoCosmosDbAccountTypeFactory" order="first"/>
+        <urlParamEditorProvider implementation="com.microsoft.azure.toolkit.intellij.cosmos.dbtools.CassandraCosmosDbAccountTypeFactory" order="first"/>
     </extensions>
 </idea-plugin>
